# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/clubs":{"post":{"tags":["Clubs"],"summary":"Create Club","description":"새로운 동아리를 생성합니다 (이미지 업로드 포함).","operationId":"create_club_clubs_post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_club_clubs_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Club"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Clubs"],"summary":"Get Clubs","description":"동아리 목록을 조회합니다.\n'name' 쿼리 파라미터가 제공되면, 해당 이름으로 동아리를 검색합니다.","operationId":"get_clubs_clubs_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Club"},"title":"Response Get Clubs Clubs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/join":{"post":{"tags":["Clubs"],"summary":"Join Club","description":"이름과 비밀번호로 특정 동아리에 참여합니다.","operationId":"join_club_clubs_join_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClubJoin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/{club_id}":{"get":{"tags":["Clubs"],"summary":"Get Club By Id","description":"ID로 특정 동아리의 상세 정보를 조회합니다.","operationId":"get_club_by_id_clubs__club_id__get","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Club"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Clubs"],"summary":"Update Club","operationId":"update_club_clubs__club_id__put","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Clubs"],"summary":"Delete Club","operationId":"delete_club_clubs__club_id__delete","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/signup":{"post":{"tags":["Authentication"],"summary":"Create User","description":"새로운 사용자를 생성(회원가입)합니다.","operationId":"create_user_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["Authentication"],"summary":"Login For Access Token","description":"사용자 로그인 후 액세스 토큰을 발급합니다.","operationId":"login_for_access_token_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/users/me":{"get":{"tags":["Authentication"],"summary":"Read Users Me","description":"현재 로그인된 사용자의 정보를 반환합니다.","operationId":"read_users_me_auth_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/clubs/{club_id}/members":{"post":{"tags":["Club Members"],"summary":"Create Member For Club","description":"특정 동아리에 새로운 부원을 추가합니다.","operationId":"create_member_for_club_clubs__club_id__members_post","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClubMemberCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClubMember"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Club Members"],"summary":"Get Members For Club","description":"특정 동아리의 모든 부원 목록을 조회합니다.","operationId":"get_members_for_club_clubs__club_id__members_get","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClubMember"},"title":"Response Get Members For Club Clubs  Club Id  Members Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/{club_id}/members/{member_id}":{"patch":{"tags":["Club Members"],"summary":"Update Member","description":"특정 부원의 정보를 수정합니다.","operationId":"update_member_clubs__club_id__members__member_id__patch","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}},{"name":"member_id","in":"path","required":true,"schema":{"type":"integer","title":"Member Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClubMemberUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClubMember"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Club Members"],"summary":"Delete Member","description":"특정 부원을 삭제합니다.","operationId":"delete_member_clubs__club_id__members__member_id__delete","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}},{"name":"member_id","in":"path","required":true,"schema":{"type":"integer","title":"Member Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/{club_id}/accounting":{"post":{"tags":["Accounting"],"summary":"Create Accounting Entry","description":"새로운 회계 내역을 생성합니다 (사진 업로드 포함).","operationId":"create_accounting_entry_clubs__club_id__accounting_post","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_accounting_entry_clubs__club_id__accounting_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountingEntry"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Accounting"],"summary":"Get Accounting Entries","description":"특정 동아리의 모든 회계 내역을 조회합니다.\nexport=true 쿼리 파라미터가 있으면 엑셀 파일로 내보냅니다.","operationId":"get_accounting_entries_clubs__club_id__accounting_get","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}},{"name":"export","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Export"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountingEntry"},"title":"Response Get Accounting Entries Clubs  Club Id  Accounting Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/{club_id}/operation-logs":{"post":{"tags":["Operation Logs"],"summary":"Create Operation Log For Club","description":"특정 동아리에 새로운 활동 기록을 생성합니다.","operationId":"create_operation_log_for_club_clubs__club_id__operation_logs_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_operation_log_for_club_clubs__club_id__operation_logs_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationLog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Operation Logs"],"summary":"Get Operation Logs For Club","description":"특정 동아리의 모든 활동 기록 목록을 조회합니다.","operationId":"get_operation_logs_for_club_clubs__club_id__operation_logs_get","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OperationLog"},"title":"Response Get Operation Logs For Club Clubs  Club Id  Operation Logs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clubs/{club_id}/operation-logs/{log_id}":{"get":{"tags":["Operation Logs"],"summary":"Get Operation Log","description":"특정 ID를 가진 활동 기록의 상세 정보를 조회합니다.","operationId":"get_operation_log_clubs__club_id__operation_logs__log_id__get","parameters":[{"name":"club_id","in":"path","required":true,"schema":{"type":"integer","title":"Club Id"}},{"name":"log_id","in":"path","required":true,"schema":{"type":"integer","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationLog"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AccountingEntry":{"properties":{"date":{"type":"string","title":"Date"},"manager":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Manager"},"description":{"type":"string","title":"Description"},"amount":{"type":"integer","title":"Amount"},"photo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Url"},"id":{"type":"integer","title":"Id"},"club_id":{"type":"integer","title":"Club Id"}},"type":"object","required":["date","description","amount","id","club_id"],"title":"AccountingEntry"},"Body_create_accounting_entry_clubs__club_id__accounting_post":{"properties":{"date":{"type":"string","title":"Date"},"description":{"type":"string","title":"Description"},"amount":{"type":"integer","title":"Amount"},"manager":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Manager"},"photo":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Photo"}},"type":"object","required":["date","description","amount"],"title":"Body_create_accounting_entry_clubs__club_id__accounting_post"},"Body_create_club_clubs_post":{"properties":{"name":{"type":"string","title":"Name"},"club_type":{"type":"string","title":"Club Type"},"topic":{"type":"string","title":"Topic"},"password":{"type":"string","title":"Password"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"file":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"File"}},"type":"object","required":["name","club_type","topic","password"],"title":"Body_create_club_clubs_post"},"Body_create_operation_log_for_club_clubs__club_id__operation_logs_post":{"properties":{"log_data":{"type":"string","title":"Log Data"},"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["log_data"],"title":"Body_create_operation_log_for_club_clubs__club_id__operation_logs_post"},"Body_login_for_access_token_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_auth_token_post"},"Club":{"properties":{"name":{"type":"string","title":"Name"},"club_type":{"type":"string","title":"Club Type"},"topic":{"type":"string","title":"Topic"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"id":{"type":"integer","title":"Id"},"members":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Members","default":[]},"club_members":{"items":{"$ref":"#/components/schemas/ClubMember"},"type":"array","title":"Club Members","default":[]},"accounting_entries":{"items":{"$ref":"#/components/schemas/AccountingEntry"},"type":"array","title":"Accounting Entries","default":[]},"operation_logs":{"items":{"$ref":"#/components/schemas/OperationLog"},"type":"array","title":"Operation Logs","default":[]}},"type":"object","required":["name","club_type","topic","id"],"title":"Club"},"ClubJoin":{"properties":{"name":{"type":"string","title":"Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","password"],"title":"ClubJoin"},"ClubMember":{"properties":{"name":{"type":"string","title":"Name"},"birth_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Date"},"student_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Id"},"major":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Major"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"member_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Member Year"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"부원"},"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"},"id":{"type":"integer","title":"Id"},"club_id":{"type":"integer","title":"Club Id"}},"type":"object","required":["name","id","club_id"],"title":"ClubMember"},"ClubMemberCreate":{"properties":{"name":{"type":"string","title":"Name"},"birth_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Date"},"student_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Id"},"major":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Major"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"member_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Member Year"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role","default":"부원"},"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"}},"type":"object","required":["name"],"title":"ClubMemberCreate"},"ClubMemberUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"birth_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Date"},"student_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Id"},"major":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Major"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"member_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Member Year"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"},"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"}},"type":"object","title":"ClubMemberUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OperationLog":{"properties":{"title":{"type":"string","title":"Title"},"post_type":{"type":"string","title":"Post Type"},"start_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"},"team":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Team"},"content":{"type":"object","title":"Content"},"id":{"type":"integer","title":"Id"},"author_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Author Id"},"files":{"items":{"$ref":"#/components/schemas/UploadedFile"},"type":"array","title":"Files","default":[]}},"type":"object","required":["title","post_type","content","id"],"title":"OperationLog"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UploadedFile":{"properties":{"id":{"type":"integer","title":"Id"},"file_name":{"type":"string","title":"File Name"},"file_path":{"type":"string","title":"File Path"}},"type":"object","required":["id","file_name","file_path"],"title":"UploadedFile"},"User":{"properties":{"email":{"type":"string","title":"Email"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"id":{"type":"integer","title":"Id"},"clubs":{"items":{"$ref":"#/components/schemas/Club"},"type":"array","title":"Clubs","default":[]},"operation_logs":{"items":{"$ref":"#/components/schemas/OperationLog"},"type":"array","title":"Operation Logs","default":[]}},"type":"object","required":["email","id"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","title":"Email"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}